requirements.txt
djangorestframework==3.11.0
Django==3.0.5
Pillow==7.1.2